Port aie_kernels to compile under xchesscc (compiler="chess")
Entire aie_kernels C++ library now builds under xchesscc and every
operator passes its reference test with compiler="chess", with the
Peano path unregressed.

Toolchain (2024 vitis_aie_essentials recipe, see CHESS_PORT_STATE.md):
- xchesscc from ~/Downloads/ryzen_ai-1.3.0/vitis_aie_essentials
  (unguarded acquire; 2025.2+/2026.1 chess guarded-acquire hangs on
  NPU2 silicon, upstream Xilinx/mlir-aie#3690)
- patched mlir-aie v1.4.2 aiecc: bare peanoElfs link (AIE_AIECC_NO_XBRIDGE),
  me_primitive stubs, self-contained aie2p softfloat (aie2p_softfloat.c),
  fixChessFloatLiterals (.0 suffix for integer-valued f32 constants),
  archive-order fix for fused OperatorSequence links
- harness: --compiler chess (conftest.py, iron/common/compilation)

Kernel fixes (peano-safe):
- bf16 scalar call-args -> float + (bfloat16) casts (chess LLVM rejects
  bf16 call constants): leaky_relu, mha partial_softmax inv_scale chain
- softmax: explicit (float) casts for comparison ambiguities
- aie2/softmax.cc: -INFINITY -> (bfloat16)(-65504.0f) (aie2 chess
  front-end headers lack INFINITY)

Test evidence (compiler="chess", 2024 essentials):
- task-1: axpy 20/20 + silu 20/20 (bare link, npu2)
- task-2: 11 elementwise/activation ops 400/400
- task-3: gemm 45 / gemv 95 / dequant 40 / mha 5 = 185/185
- task-4: transpose/strided_copy/repeat/mem_copy 140/140 + axpy 20/20
- task-5: Qwen3-0.6B smoke_layer LAYER_OK + HEAD_OK under chess and peano
- rope/swiglu reference tests: chess 55 passed, peano 55 passed
- peano regression clean across all suites; clean-state re-runs verified

+++ aie_kernels/aie2p/mha.cc
@@ -60,7 +60,7 @@ void partial_softmax_bf16(bfloat16 *input,
                           const int32_t input_size,
                           const int32_t row_idx,
                           const int32_t row_size,
-                          const bfloat16 scale);
+                          const float scale);
 void passThroughLine(int32_t *in, int32_t *out, int32_t lineWidth);
 
 void matmul_bf16_bf16_wrapper(bfloat16 *a_in, bfloat16 *b_in, bfloat16 *c_out, int32_t *idx_buffer)
@@ -164,7 +164,7 @@ void partial_softmax(bfloat16 *A,
                      bfloat16 *P,
                      bfloat16 *scale_buffer,
                      int32_t *idx_buffer,
-                     bfloat16 inv_scale,
+                     float inv_scale,
                      int32_t B_q,
                      int32_t B_kv,
                      int32_t S_q_eff,
